pyproject.toml
[tool.poetry]
name = "PyGT521F32"
version = "0.1.3"
description = "An SDK module for the ADH-Tech GT521F32 fingerprint scanner"
authors = ["Alon <alonlivne@gmail.com>"]
packages = [
	{ include = "GT521F32"},
	{ include = "GT521F32-viewer"},
]
	

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.4"
Pillow = "^7.2.0"
cython-sgio = {version = "^1.1.2", platform = "linux"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
